cleared database and migrations and added intital data
GO JOSÈ
