+++ GirafRest/Data/DBInitializer.cs
@@ -10,8 +10,9 @@ using System.Drawing.Imaging;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
+using GirafRest.Data.Samples;
 
-namespace GirafRest.Setup
+namespace GirafRest.Data
 {
     /// <summary>
     /// A class for initializing the database with some sample data.
@@ -34,7 +35,7 @@ namespace GirafRest.Setup
         /// <param name="pictogramCount">The number of sample pictograms to generate.</param>
         public static async Task Initialize(GirafDbContext context, UserManager<GirafUser> userManager, int pictogramCount)
         {
-            CreatePictograms(pictogramCount);
+            context.Database.EnsureCreated();
 
             // Check if any data is in the database
             if (await context.Departments.AnyAsync())
@@ -48,6 +49,9 @@ namespace GirafRest.Setup
                 //await SampleDataHandler.SerializeDataAsync(context, userManager);
                 return;
             }
+
+            CreatePictograms(pictogramCount);
+
             // Get sample data
             SampleData sampleData = SampleDataHandler.DeserializeData();
             // Create departments
@@ -74,7 +78,7 @@ namespace GirafRest.Setup
                 }
             }
             // Save changes
-            await context.SaveChangesAsync();
+            await context.SaveChangesAsync();   
         }
 
         #region Generating pictograms
