+++ GirafRest/Data/SampleData.cs
@@ -1,0 +1,29 @@
+﻿using GirafRest.Data.Samples;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace GirafRest.Data
+{
+    public class SampleData
+    {
+        public List<SampleGirafUser> UserList { get; set; }
+        public List<SampleDepartment> DepartmentList { get; set; }
+        public List<SamplePictogram> PictogramList { get; set; }
+        public List<SampleWeekday> WeekdayList { get; set; }
+        public List<SampleWeek> WeekList { get; set; }
+        public List<SampleWeekTemplate> WeekTemplateList { get; set; }
+
+
+        public SampleData()
+        {
+            UserList = new List<SampleGirafUser>();
+            DepartmentList = new List<SampleDepartment>();
+            PictogramList = new List<SamplePictogram>();
+            WeekdayList = new List<SampleWeekday>();
+            WeekList = new List<SampleWeek>();
+            WeekTemplateList = new List<SampleWeekTemplate>();
+        }
+    }
+}
