+++ GirafRest/Data/Samples/SampleGirafUser.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace GirafRest.Setup
+namespace GirafRest.Data.Samples
 {
     public class SampleGirafUser
     {
