+++ GirafRest/Program.cs
@@ -5,6 +5,10 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using MySql.Data.MySqlClient;
 using System;
+using GirafRest.Data;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.AspNetCore.Identity;
+using GirafRest.Models;
 
 namespace GirafRest
 {
@@ -36,6 +40,7 @@ namespace GirafRest
             {
                 BuildWebHost(args).Run();
                 Console.WriteLine("webhost built");
+                
             }
             catch (MySqlException e)
             {
@@ -52,8 +57,9 @@ namespace GirafRest
         /// <see cref="Startup"/> sets the general environment (authentication, logging i.e)
         /// </summary>
         /// <returns>A <see cref="IWebHost"/> host fit for running the server.</returns>
-        public static IWebHost BuildWebHost(string[] args) =>
-        WebHost.CreateDefaultBuilder()
+        public static IWebHost BuildWebHost(string[] args)
+        {
+          return WebHost.CreateDefaultBuilder()
                .UseKestrel()
                .UseUrls($"http://+:{ProgramOptions.Port}")
                .UseIISIntegration()
@@ -68,5 +74,8 @@ namespace GirafRest
                 })
                .UseDefaultServiceProvider(options => options.ValidateScopes = false)
                .Build();
+            
+        }
+       
     }
 }
